fix(guild): DismissGuildMemberAndNotice 完整实现 + 删除多余 uchar 重载
- 原版遍历成员发 Dismiss 通知包 + Guild_Exp_Book_Delete 包 + ResetGuild
  （我们此前只有 DetachGuild）
- 原版仅 int 重载，删除无符号 char 桩重载并适配调用点

+++ dnf_decompile/source/guild/GuildPacket.cpp
@@ -394,7 +394,7 @@ void CPacketTranslater::OnNoticeGuildDismiss(PacketHeader* pkt)
     CGuild* guild = m_pclApp->Get_GuildManager()->FindGuild(*(unsigned int*)(pb + 10));
     if (guild != 0)
     {
-        guild->DismissGuildMemberAndNotice(m_pclApp->Get_ServerGroup());
+        guild->DismissGuildMemberAndNotice((int)m_pclApp->Get_ServerGroup());
         m_pclApp->Get_GuildManager()->GuildDismiss(guild);
     }
 }
